refactor(dashboard): delete unnecessary 'EmptyList' component
