+++ src/containers/Devices/components/DevicesList.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import ReactWinJS from 'react-winjs'
 import WinJS from 'winjs'
 import DevicesItemList from './DevicesItemList'
-import BuildItemList from '../../../components/BuildItemList'
+import BuildItemList from '../../../shared/BuildItemList'
 import Loader from '../../../components/Loader'
 import Confirmation from '../../../components/Confirmation'
 import EmptyMessage from '../../../components/EmptyMessage'
