+++ src/containers/Invitations/components/InvitationsList.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import ReactWinJS from 'react-winjs'
 import InvitationsItemList from './InvitationsItemList'
-import BuildItemList from '../../../components/BuildItemList'
+import BuildItemList from '../../../shared/BuildItemList'
 import WinJS from 'winjs'
 import Loader from '../../../components/Loader'
 import Confirmation from '../../../components/Confirmation'
